experiment with explicitly adding conscrypt to SBOM
spark-connect can use concrypt if it is on the classpath, but
doesn't depend on it directly. It would be interesting to see
if that fact could be encoded in the SBOM.

Create the SBOM with:

```
mvn -pl connector/connect/server cyclonedx:makeBom
```

And then view it at:

```
connector/connect/server/target/bom.json
```

It currently does not refer to Conscrypt.